improved tooltips, tree added successfully
